Fix: Deprecated use of last arg. as keyword param
As of Ruby 2.7 the use of hashes in the last argument of a method
call has been deprecated and results in a warning.
Since Ruby automatically converts the last argument in a method to
keyword arguments if it's a hash, simply removing the curly braces
of the hash argument will be enough in this case.